move isEncryptedData back to util.test
