+++ packages/sdk/src/types.ts
@@ -115,23 +115,6 @@ export function isConfirmedEvent(event: StreamTimelineEvent): event is Confirmed
     )
 }
 
-// Type guard function based on field checks
-export function isEncryptedData(obj: unknown): obj is EncryptedData {
-    if (typeof obj !== 'object' || obj === null) {
-        return false
-    }
-
-    const data = obj as EncryptedData
-    return (
-        typeof data.ciphertext === 'string' &&
-        typeof data.algorithm === 'string' &&
-        typeof data.senderKey === 'string' &&
-        typeof data.sessionId === 'string' &&
-        (typeof data.checksum === 'string' || data.checksum === undefined) &&
-        (typeof data.refEventId === 'string' || data.refEventId === undefined)
-    )
-}
-
 export function makeRemoteTimelineEvent(params: {
     parsedEvent: ParsedEvent
     eventNum: bigint
