+++ packages/sdk/src/util.test.ts
@@ -754,3 +754,20 @@ export async function createChannel(
     }
     return { channelId, error: undefined }
 }
+
+// Type guard function based on field checks
+export function isEncryptedData(obj: unknown): obj is EncryptedData {
+    if (typeof obj !== 'object' || obj === null) {
+        return false
+    }
+
+    const data = obj as EncryptedData
+    return (
+        typeof data.ciphertext === 'string' &&
+        typeof data.algorithm === 'string' &&
+        typeof data.senderKey === 'string' &&
+        typeof data.sessionId === 'string' &&
+        (typeof data.checksum === 'string' || data.checksum === undefined) &&
+        (typeof data.refEventId === 'string' || data.refEventId === undefined)
+    )
+}
